// src/mortgage/spain/domain/types.ts
import { OtherFees } from "../../domain/types";

export interface MortgageRules {
    loanConstraints: LoanConstraints;
    interest: InterestRules;
    transferTax: TransferTaxRules;
}

export interface LoanConstraints {
    maxLtvPercent: number;
    maxAmortizationYears: number;
}

export interface InterestRules {
    compounding: 'MONTHLY';
}

export interface TransferTaxBracket {
    upTo?: number;
    above?: number;
    rate: number;
}

export interface TransferTaxRules {
    brackets: TransferTaxBracket[];
}

export interface AmortizationScheduleItem {
    year: number;
    principal: number;
    interest: number;
    balance: number;
}

export interface MortgageInput {
    propertyPrice: number;
    downPayment: number;
    annualInterestRate: number;
    amortizationYears: number;
    isFirstTimeBuyer: boolean;
}

export interface MortgageOutput {
    loanAmount: number;
    totalMortgage: number;
    monthlyPayment: number;
    totalInterestPaid: number;
    totalPaid: number;
    transferTax: number;
    amortizationSchedule: AmortizationScheduleItem[];
    otherFees: OtherFees;
}